outdated/proto_pico/requirements.txt
pyserial
dynamixel-sdk
PyYAML
